aftercare_voice_agent/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
python-dotenv>=1.0
requests>=2.32
elevenlabs>=1.50
